Split RichState file
